feat(lib): delete swagger from everywhere!
