+++ lib/power_api/engine.rb
@@ -15,7 +15,6 @@ module PowerApi
       require_relative "./generator_helper/active_record_resource"
       require_relative "./generator_helper/api_helper"
       require_relative "./generator_helper/resource_helper"
-      require_relative "./generator_helper/swagger_helper"
       require_relative "./generator_helper/ams_helper"
       require_relative "./generator_helper/rspec_controller_helper"
       require_relative "./generator_helper/controller_helper"
