+++ lib/power_api/generator_helpers.rb
@@ -3,7 +3,6 @@ module PowerApi
     include GeneratorHelper::ControllerActionsHelper
     include GeneratorHelper::ResourceHelper
     include GeneratorHelper::ApiHelper
-    include GeneratorHelper::SwaggerHelper
     include GeneratorHelper::RspecControllerHelper
     include GeneratorHelper::AmsHelper
     include GeneratorHelper::ControllerHelper
